+++ assets/js/views/main.js
@@ -4,12 +4,22 @@ import 'bootstrap'
 
 export default class MainView {
 
+  // reload page preserving scroll position and preventing the creation of a new history entry
+  // https://github.com/turbolinks/turbolinks/issues/329#issuecomment-341699031
+  reload() {
+    localStorage.setItem('scroll-pos', $(window).scrollTop());
+    Turbolinks.visit(window.location, { action: 'replace' })
+  }
+
   // This will be executed whenever a new page is loaded
   mount() {
 
-    // scroll back to previous scroll position on page load
-    const pos = localStorage.getItem('scroll-pos', 0);
-    if (pos) $(window).scrollTop(pos)
+    // load saved scroll position and reset
+    const pos = localStorage.getItem('scroll-pos');
+    if (pos){
+      $(window).scrollTop(pos)
+      localStorage.removeItem('scroll-pos');
+    }
 
     // enable tooltips
     $('[data-toggle="tooltip"]').tooltip();
@@ -55,8 +65,8 @@ export default class MainView {
     //   style_formats_merge: true
     // });
 
-    // enable form validation removal on change
-    $('.form-control').change((e) => {
+    // enable form validation removal on change for all bootstrap input fields
+    $('body').on('keyup', '.form-control', (e) => {
       $(e.target)
         .removeClass('is-valid is-invalid')
         .siblings('.invalid-feedback').remove();
@@ -66,10 +76,7 @@ export default class MainView {
   unmount() {
     // This will be executed when the document unloads...
 
-    // reset scroll position
-    localStorage.setItem('scroll-pos', 0);
-
     clearInterval(this.timer);
-    $( ".form-control" ).off();
+    $('body').off();
   }
 }